requirements.txt
piper-tts>=1.2.0
numpy>=1.24.0
scipy>=1.10.0
pydub>=0.25.0
requests>=2.28.0